Enable the service that creates the file share symlink
This service is a systemd user-mode service for the vnc user.  Note
that we have to enable linger for the vnc user before enabling the
service, and we disable it afterwards.  This is just a trick to get
the vnc user's user-specific systemd session to start without actually
logging in as that user.